+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 
-All notable changes to the Line Width Indicator extension will be documented in this file.
+All notable changes to the [Line Width Indicator](https://github.com/lbragile/Line-Width-Indicator) extension will be documented in this file.
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
+
+## [0.0.5] - 3/30/2021
+
+### Added
+
+- Automatic comment insertion to prevent formatting of the line with settings to enable/disable.
 
 ## [0.0.4] - 3/29/2021